WIP on auth

+++ src/main/java/nz/co/mircle/v1/api/user/controller/UserController.java
@@ -15,6 +15,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+import java.util.List;
+
 /** Here are a lists of user API. */
 @RestController
 @Api(value = "user", description = "User API")
@@ -91,6 +94,70 @@ public class UserController extends AbstractController {
     return new ResponseEntity<>(user, HttpStatus.CREATED);
   }
 
+  @ApiOperation(value = "Getting a user by id", response = Iterable.class)
+  @ApiResponses(
+          value = {
+                  @ApiResponse(code = 200, message = "Successfully retrieved a user"),
+                  @ApiResponse(code = 201, message = "Successfully retrieved a user"),
+                  @ApiResponse(code = 401, message = "You are not authorized to retrieved a user."),
+                  @ApiResponse(
+                          code = 403,
+                          message = "Accessing the resource you were trying to reach is forbidden"
+                  ),
+                  @ApiResponse(code = 404, message = "The resource you were trying to reach is not found")
+          }
+  )
+  @RequestMapping(value = "/{id}/friend", method = RequestMethod.POST)
+  public ResponseEntity addFriend(@PathVariable("id") Long id, Principal principal) {
+    LOG.info(String.format("Adding user ID %d friend ID...", id));
+
+    User currentUser = userService.findUser(principal.getName());
+    try {
+      if (currentUser.getId() == id) {
+        userService.addFriend(id);
+        LOG.info("User %d friends found.");
+      } else {
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+      }
+    } catch (Exception e) {
+      LOG.error(String.format("Attempt to find a user with id %d friends failed.", id));
+      LOG.error(e.getMessage());
+      return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    return new ResponseEntity<>(HttpStatus.OK);
+  }
+
+  @ApiOperation(value = "Getting a user by id", response = Iterable.class)
+  @ApiResponses(
+          value = {
+                  @ApiResponse(code = 200, message = "Successfully retrieved a user"),
+                  @ApiResponse(code = 201, message = "Successfully retrieved a user"),
+                  @ApiResponse(code = 401, message = "You are not authorized to retrieved a user."),
+                  @ApiResponse(
+                          code = 403,
+                          message = "Accessing the resource you were trying to reach is forbidden"
+                  ),
+                  @ApiResponse(code = 404, message = "The resource you were trying to reach is not found")
+          }
+  )
+  @RequestMapping(value = "/{id}/friends", method = RequestMethod.GET)
+  public ResponseEntity findFriendsById(@PathVariable("id") Long id) {
+    LOG.info(String.format("Getting user ID %d friends...", id));
+
+    List<User> friends;
+    try {
+      friends = userService.findFriends(id);
+      LOG.info("User %d friends found.");
+    } catch (Exception e) {
+      LOG.error(String.format("Attempt to find a user with id %d friends failed.", id));
+      LOG.error(e.getMessage());
+      return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    return new ResponseEntity<>(friends, HttpStatus.CREATED);
+  }
+
   @ApiOperation(value = "Delete a user", response = Iterable.class)
   @ApiResponses(
           value = {
@@ -105,12 +172,17 @@ public class UserController extends AbstractController {
           }
   )
   @RequestMapping(method = RequestMethod.DELETE)
-  public ResponseEntity deleteUser(@PathVariable("id") Long id) {
+  public ResponseEntity deleteUser(@PathVariable("id") Long id, Principal principal) {
     LOG.info(String.format("Deleting user with id %s...", id));
 
+    User currentUser = userService.findUser(principal.getName());
     try {
-      userService.deleteUser(id);
-      LOG.info(String.format("User %d deleted.", id));
+      if (currentUser.getId() == id) {
+        userService.deleteUser(id);
+        LOG.info(String.format("User %d deleted.", id));
+      } else {
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+      }
     } catch (Exception e) {
       LOG.error(String.format("Attempt to delete user with id %d failed.", id));
       LOG.error(e.getMessage());
